# trading-service/controllers/SecuritiesController.go
package controllers

import (
	"banka1.com/db"
	"banka1.com/dto"
	"banka1.com/types"
	"github.com/gofiber/fiber/v2"
)

type SecuritiesController struct {
}

func NewSecuritiesController() *SecuritiesController {
	return &SecuritiesController{}
}

// GetAllSecurities godoc
//
//	@Summary		Preuzimanje svih hartija od vrednosti (Alias)
//	@Description	Vraća listu svih dostupnih hartija od vrednosti.
//	@Tags			Securities
//	@Produce		json
//	@Success		200	{object}	types.Response{data=[]types.Security}	"Lista svih hartija od vrednosti"
//	@Failure		500	{object}	types.Response							"Interna greška servera pri preuzimanju ili konverziji hartija od vrednosti"
//	@Router			/securities [get]
func (sc *SecuritiesController) getSecurities() func(c *fiber.Ctx) error {
	return func(c *fiber.Ctx) error {
		var listings []types.Listing
		if result := db.DB.Preload("Exchange").Find(&listings); result.Error != nil {
			return c.Status(500).JSON(types.Response{
				Success: false,
				Data:    nil,
				Error:   "Failed to fetch securities: " + result.Error.Error(),
			})
		}
		var securities []types.Security
		for _, listing := range listings {
			security, err := listingToSecurity(&listing)
			if err != nil || security == nil {
				return c.Status(500).JSON(types.Response{
					Success: false,
					Data:    nil,
					Error:   "Failed to convert listing to security: " + err.Error(),
				})
			}
			securities = append(securities, *security)
		}

		return c.JSON(types.Response{
			Success: true,
			Data:    securities,
			Error:   "",
		})
	}
}

// GetUserSecurities godoc
//
//	@Summary		Dohvatanje svih hartija od vrednosti za korisnika
//	@Description	Vraća listu svih hartija od vrednosti koje poseduje korisnik sa specificiranim ID-jem, zajedno sa količinom, kupovnom cenom i trenutnim profitom/gubitkom za svaku.
//	@Tags			Securities
//	@Produce		json
//	@Param			id	path		string											true	"ID korisnika čije hartije od vrednosti treba dohvatiti"	example(123)
//	@Success		200	{object}	types.Response{data=[]dto.PortfolioSecurityDTO}	"Lista hartija od vrednosti u vlasništvu korisnika"
//	@Failure		400	{object}	types.Response									"ID korisnika nije prosleđen u putanji"
//	@Failure		500	{object}	types.Response									"Greška pri dohvatanju podataka iz baze"
//	@Router			/securities/{id} [get]
func (sc *SecuritiesController) GetUserSecurities(c *fiber.Ctx) error {
	userId := c.Params("id")
	if userId == "" {
		return c.Status(400).JSON(types.Response{
			Success: false,
			Error:   "Nedostaje user ID",
		})
	}

	var portfolios []types.Portfolio
	if err := db.DB.Preload("Security").Where("user_id = ?", userId).Find(&portfolios).Error; err != nil {
		return c.Status(500).JSON(types.Response{
			Success: false,
			Error:   "Greška pri dohvatanju portfolija: " + err.Error(),
		})
	}

	var response []dto.PortfolioSecurityDTO
	for _, p := range portfolios {
		profit := (p.Security.LastPrice - p.PurchasePrice) * float64(p.Quantity)
		item := dto.PortfolioSecurityDTO{
			Ticker:       p.Security.Ticker,
			Type:         p.Security.Type,
			Amount:       p.Quantity,
			Price:        p.PurchasePrice,
			Profit:       profit,
			LastModified: p.CreatedAt,
		}
		response = append(response, item)
	}

	return c.JSON(types.Response{
		Success: true,
		Data:    response,
		Error:   "",
	})
}

func listingToSecurity(l *types.Listing) (*types.Security, error) {
	var security types.Security
	switch l.Type {
	case "Stock":
		{
			security = types.Security{
				ID:        l.ID,
				Ticker:    l.Ticker,
				Name:      l.Name,
				Type:      l.Type,
				Exchange:  l.Exchange.Name,
				LastPrice: float64(l.Price),
				AskPrice:  float64(l.Ask),
				BidPrice:  float64(l.Bid),
				Volume:    int64(l.ContractSize * 10),
			}
		}
	case "Forex":
		{
			security = types.Security{
				ID:        l.ID,
				Ticker:    l.Ticker,
				Name:      l.Name,
				Type:      l.Type,
				Exchange:  l.Exchange.Name,
				LastPrice: float64(l.Price),
				AskPrice:  float64(l.Ask),
				BidPrice:  float64(l.Bid),
				Volume:    int64(l.ContractSize * 10),
			}
		}
	case "Future":
		{
			var future types.FuturesContract
			if result := db.DB.Where("listing_id = ?", l.ID).First(&future); result.Error != nil {
				return nil, result.Error
			}
			settlementDate := future.SettlementDate.Format("2006-01-02")
			security = types.Security{
				ID:             l.ID,
				Ticker:         l.Ticker,
				Name:           l.Name,
				Type:           l.Type,
				Exchange:       l.Exchange.Name,
				LastPrice:      float64(l.Price),
				AskPrice:       float64(l.Ask),
				BidPrice:       float64(l.Bid),
				Volume:         int64(l.ContractSize * 10),
				SettlementDate: &settlementDate,
			}
		}
	case "Option":
		{
			var option types.Option
			if result := db.DB.Where("listing_id = ?", l.ID).First(&option); result.Error != nil {
				return nil, result.Error
			}
			security = types.Security{
				ID:             l.ID,
				Ticker:         l.Ticker,
				Name:           l.Name,
				Type:           l.Type,
				Exchange:       l.Exchange.Name,
				LastPrice:      float64(l.Price),
				AskPrice:       float64(l.Ask),
				BidPrice:       float64(l.Bid),
				Volume:         int64(l.ContractSize * 10),
				StrikePrice:    &option.StrikePrice,
				OptionType:     &option.OptionType,
				SettlementDate: nil,
			}

		}

	}
	return &security, nil
}

func InitSecuritiesRoutes(app *fiber.App) {
	securitiesController := NewSecuritiesController()

	app.Get("/securities", securitiesController.getSecurities())
	app.Get("/securities/available", securitiesController.getSecurities())
	app.Get("/securities/:id", securitiesController.GetUserSecurities)
}