# trading-service/docs/swagger.yaml
basePath: /
definitions:
...
  dto.PortfolioSecurityDTO:
    properties:
      amount:
        type: integer
      last_modified:
        type: integer
      price:
        type: number
      profit:
        type: number
      ticker:
        type: string
      type:
        type: string
    type: object
...
  types.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  types.Security:
    properties:
      ask:
        type: number
      availableQuantity:
        type: integer
      bid:
        type: number
      exchange:
        type: string
      id:
        type: integer
      lastPrice:
        type: number
      name:
        type: string
      optionType:
        type: string
      settlementDate:
        description: Samo za futures i opcije
        type: string
      strikePrice:
        description: data for option
        type: number
      ticker:
        type: string
      type:
        type: string
    type: object
...
  types.TaxResponse:
    properties:
      is_actuary:
        type: boolean
      is_paid:
        type: boolean
      tax_amount:
        type: number
      taxable_profit:
        type: number
      user_id:
        type: integer
    type: object
host: localhost:3000
info:
  contact: {}
  description: Trading Service API
  title: Trading Service
  version: "1.0"
paths:
...
  /securities:
    get:
      description: Vraća listu svih dostupnih hartija od vrednosti.
      produces:
      - application/json
      responses:
        "200":
          description: Lista svih hartija od vrednosti
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Security'
                  type: array
              type: object
        "500":
          description: Interna greška servera pri preuzimanju ili konverziji hartija
            od vrednosti
          schema:
            $ref: '#/definitions/types.Response'
      summary: Preuzimanje svih hartija od vrednosti (Alias)
      tags:
      - Securities
  /securities/{id}:
    get:
      description: Vraća listu svih hartija od vrednosti koje poseduje korisnik sa
        specificiranim ID-jem, zajedno sa količinom, kupovnom cenom i trenutnim profitom/gubitkom
        za svaku.
      parameters:
      - description: ID korisnika čije hartije od vrednosti treba dohvatiti
        example: "123"
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista hartija od vrednosti u vlasništvu korisnika
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PortfolioSecurityDTO'
                  type: array
              type: object
        "400":
          description: ID korisnika nije prosleđen u putanji
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Greška pri dohvatanju podataka iz baze
          schema:
            $ref: '#/definitions/types.Response'
      summary: Dohvatanje svih hartija od vrednosti za korisnika
      tags:
      - Securities
...
  /tax:
    get:
      description: Vraća listu najskorijih neplaćenih poreskih obaveza (za poslednji
        obračunati mesec/godinu) za sve korisnike. Za svakog korisnika proverava i
        da li je registrovan kao aktuar.
      produces:
      - application/json
      responses:
        "200":
          description: Lista poslednjih neplaćenih poreskih obaveza
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.TaxResponse'
                  type: array
              type: object
        "400":
          description: Greška pri izvršavanju upita u bazi (kako je implementirano
            u kodu)
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Greška pri čitanju rezultata iz baze
          schema:
            $ref: '#/definitions/types.Response'
      summary: Dohvatanje poslednjeg neplaćenog poreza za sve korisnike
      tags:
      - Tax
  /tax/run:
    post:
      description: Endpoint namenjen za pokretanje procesa obračuna poreza za korisnike.
        Trenutno nije implementiran i uvek vraća grešku 500.
      produces:
      - application/json
      responses:
        "202":
          description: Zahtev za obračun poreza je primljen
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Greska
          schema:
            $ref: '#/definitions/types.Response'
      summary: Pokretanje obračuna poreza
      tags:
      - Tax
securityDefinitions:
  BearerAuth:
    description: 'Unesite token. Primer: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"